requirements.txt
google-generativeai
python-dotenv
Pillow
pandas
scikit-learn
lightgbm
pytest